Found and fixed a logic bug
My hosting company moved my accounts to all new servers.  The system
time on the web server was 8 seconds ahead of the time on the MySQL
server.  This exposed a bug in how the timestamps worked in the
half-hourly temperature logging.
Assume webserver system time is 02:00:00.  This makes the MySQL server
time 01:59:52.  So cron on the web server launched the process and the
timestamp is computed form MySQL and the HH:MM:SS is forced to conform
to HH:M0:00 so that the charting code will pick it up.  But that time is
01:50:00 giving the appearance of firing ten minutes early AND
guaranteeing that it will NOT be picked up by the chart.

Because cron does not use the .profile a new config file for shell
scripts is introduced: config.ksh
The cron launch scripts now reference that config file.  With additional
new code to make sure the right directory is used without the need for
hard coding.
The actual modification to the temperature log for this bug is small,
but the previous communication library changes are now taken advantage
of (so this also somewhat deals with issue #16)
Other files have code cleanup and additional comments added.

+++ scripts/thermo_update_temps.php
@@ -2,30 +2,44 @@
 $start_time = microtime(true);
 require(dirname(__FILE__).'/../common.php');
 
-$log->logInfo( 'temps: start' );
+$log->logInfo( 'temps: Start.' );
 $today = date( 'Y-m-d' );
 $yesterday = date( 'Y-m-d', strtotime( 'yesterday' ));
+if( $argc < 2 )
+{
+	$log->logError( 'temps: required argument missing.  Must send unix timestamp!' );
+	die();
+}
+$unixTime = $argv[1];	// argv[0] is this files name
 
 /**
-	* This script updates the indoor and outdoor temperatures and today's and yesterday total run time for each thermostat
+	* This script updates the indoor and outdoor temperatures and today's and yesterday total run time for each thermostat.
 	*/
-
 try
 {
-	$sql = "INSERT INTO {$dbConfig['table_prefix']}temperatures( tstat_uuid, date, indoor_temp, outdoor_temp, indoor_humidity, outdoor_humidity ) VALUES ( ?, CONCAT( SUBSTR( NOW() , 1, 15 ) , \"0:00\" ), ?, ?, ?, ? )";
-	$queryTemp = $pdo->prepare($sql);
+	$sql = "SELECT NOW() as now_time, CONCAT( SUBSTR( NOW() , 1, 15 ) , '0:00' ) as magic_time;";
+	$queryMySQLServer = $pdo->prepare( $sql );
+
+	//$sql = "INSERT INTO {$dbConfig['table_prefix']}temperatures( tstat_uuid, date, indoor_temp, outdoor_temp, indoor_humidity, outdoor_humidity ) VALUES ( ?, CONCAT( SUBSTR( NOW() , 1, 15 ) , \"0:00\" ), ?, ?, ?, ? )";
+	$sql = "INSERT INTO {$dbConfig['table_prefix']}temperatures( tstat_uuid, date, indoor_temp, outdoor_temp, indoor_humidity, outdoor_humidity ) VALUES ( ?, $unixTime, ?, ?, ?, ? )";
+	$queryTemp = $pdo->prepare( $sql );
 
 	$sql = "DELETE FROM {$dbConfig['table_prefix']}run_times WHERE date = ? AND tstat_uuid = ?";
-	$queryRunDelete = $pdo->prepare($sql);
+	$queryRunDelete = $pdo->prepare( $sql );
 
 	$sql = "INSERT INTO {$dbConfig['table_prefix']}run_times( tstat_uuid, date, heat_runtime, cool_runtime ) VALUES ( ?, ?, ?, ? )";
-	$queryRunInsert = $pdo->prepare($sql);
+	$queryRunInsert = $pdo->prepare( $sql );
 }
 catch( Exception $e )
 {
-	$log->logInfo( 'temps: DB Exception while preparing SQL: ' . $e->getMessage() );
+	$log->logError( 'temps: DB Exception while preparing SQL: ' . $e->getMessage() );
 	die();
 }
+
+$queryMySQLServer->execute();
+$row = $queryMySQLServer->fetch( PDO::FETCH_ASSOC );
+$log->logInfo( "temps: The MySQL server thinks that the magic formatted time is {$row['magic_time']} where unix (on the webserver) thinks it is $unixTime" );
+
 
 $outdoorTemp = null;						// Default outside temp
 $outdoorHumidity = null;				// Default outside humidity (in case not working with CT80 or similar unit)
@@ -35,11 +49,12 @@ try
 	$outsideData = $externalWeatherAPI->getOutdoorWeather( $ZIP );
 	$outdoorTemp = $outsideData['temp'];
 	$outdoorHumidity = $outsideData['humidity'];
-	$log->logInfo( "temps: Outside Weather for {$ZIP}: Temp $outdoorTemp Humidity $outdoorHumidity" );
+//$log->logInfo( "temps: Outside Weather for {$ZIP}: Temp $outdoorTemp Humidity $outdoorHumidity" );
 }
 catch( Exception $e )
 {
-	$log->logInfo( 'temps: External weather failed: ' . $e->getMessage() );
+	$log->logError( 'temps: External weather failed: ' . $e->getMessage() );
+	// Not a fatal error, keep going.
 }
 
 foreach( $thermostats as $thermostatRec )
@@ -48,7 +63,7 @@ foreach( $thermostats as $thermostatRec )
 	$lock = @fopen( $lockFileName, 'w' );
 	if( !$lock )
 	{
-		$log->logInfo( "temps: Could not write to lock file $lockFileName" );
+		$log->logError( "temps: Could not write to lock file $lockFileName" );
 		continue;
 	}
 
@@ -58,52 +73,61 @@ foreach( $thermostats as $thermostatRec )
 		{
 			// Query thermostat info
 			$indoorHumidity = null;
-			$log->logInfo( "temps: Connecting to {$thermostatRec['id']} {$thermostatRec['tstat_uuid']} {$thermostatRec['ip']} {$thermostatRec['name']}" );
+//$log->logInfo( "temps: Connecting to {$thermostatRec['id']} {$thermostatRec['tstat_uuid']} {$thermostatRec['ip']} {$thermostatRec['name']}" );
 			$stat = new Stat( $thermostatRec['ip'] );
 
-			//$sysInfo = $stat->getSysInfo();
-			$stat->getSysInfo();
-$log->logInfo( "temps: Back from low level communication I have the error code as ($stat->connectOK)" );
-			//$uuid = $sysInfo['uuid'];
-			//$model = $stat->getModel();
-			$stat->getModel();
-			//$statData = $stat->getStat();
-			$stat->getStat();
-
-
-// Instead of asking the thermostat what his model is, rely upon the entry in the thermostat table
-			if( strstr($stat->model, 'CT80') !== false )
-			{ // Get indoor humidity for CT80
-				//sleep(2); // let thermostat catch up
-				//$indoorHumidity = $stat->getHumidity();
-				$stat->getHumidity();
-// Actually, won't this have come back from the getStat() call already if it's available????
+			$stat->getSysInfo();	// Get uuid for for insert key (yuck)
+			if( $stat->connectOK != 0 )
+			{
+				$log->logError( "temps: Error getting UUID from {$thermostatRec['id']} {$thermostatRec['tstat_uuid']} {$thermostatRec['ip']} {$thermostatRec['name']}.  Aborting communication with this unit." );
+				continue;
 			}
 
-			// Log the indoor and outdoor temperatures for this half-hour increment
+			$stat->getModel();		// Get model to know if humidity is available
+// Instead of asking the thermostat what his model is, should rely upon the entry in the thermostat table?
+			if( strstr($stat->model, 'CT80') !== false )
+			{ // Get indoor humidity for CT80
+				$stat->getHumidity();
+// Actually, won't the humidity come back from the getStat() call if it is available on the thermostat?
+			}
 
-//			$log->logInfo( "temps: UUID $stat->uuid IT " . $stat->temp . " OT $outdoorTemp IH $stat->humidity OH $outdoorHumidity" );
-			$queryTemp->execute(array( $stat->uuid, $stat->temp, $outdoorTemp, $stat->humidity, $outdoorHumidity ) );
+			// Fetch and log the indoor and outdoor temperatures for this half-hour increment
+			$stat->getStat();
+//$log->logInfo( "temps: Back from low level communication I have the error code as ($stat->connectOK)" );
+//$log->logInfo( "temps: Back from low level communication I have the temperature as ($stat->temp)" );
+//$log->logInfo( "temps: UUID $stat->uuid IT " . $stat->temp . " OT $outdoorTemp IH $stat->humidity OH $outdoorHumidity  at PHP time = " . date("Y-m-d H:i:s") );
+			if( $stat->connectOK == 0 )
+			{
+				$queryTemp->execute(array( $stat->uuid, $stat->temp, $outdoorTemp, $stat->humidity, $outdoorHumidity ) );
+			}
+			else
+			{
+				$log->logError( "temps: Error getting temperatures from {$thermostatRec['id']} {$thermostatRec['tstat_uuid']} {$thermostatRec['ip']} {$thermostatRec['name']}.  No data stored." );
+			}
 
-			//$runTimeData = $stat->getDataLog();
+
+			// Fetch and log the run time for yesterday and today
 			$stat->getDataLog();
-
-			// Need to verify success of thermostat query before deleting/inserting data...
-			// Unless it throws an exception?
-
-			// Remove zero or one rows for today and then insert one row for today.
-			$queryRunDelete->execute( array($today, $stat->uuid) );
 //			$log->logInfo( "temps: Run Time Today - Inserting RTH {$stat->runTimeHeat} RTC {$stat->runTimeCool} U $stat->uuid T $today" );
-			$queryRunInsert->execute( array($stat->uuid, $today, $stat->runTimeHeat, $stat->runTimeCool) );
-
-			// Remove zero or one rows for yesterday and then insert one row for yesterday.
-// Ought to keep track of when "yesterday" was last updated and if it was any time "today" then skip this!
-			$queryRunDelete->execute( array($yesterday, $stat->uuid) );
 //			$log->logInfo( "temps: Run Time Yesterday - Inserting RTH {$stat->runTimeHeatYesterday} RTC {$stat->runTimeCoolYesterday} U $stat->uuid T $yesterday" );
-			$queryRunInsert->execute( array($stat->uuid, $yesterday, $stat->runTimeHeatYesterday, $stat->runTimeCoolYesterday) );
+
+			if( $stat->connectOK == 0 )
+			{
+				$queryRunDelete->execute( array($today, $stat->uuid) );	// Remove zero or one rows for today and then insert one row for today.
+				$queryRunInsert->execute( array($stat->uuid, $today, $stat->runTimeHeat, $stat->runTimeCool) );	// Add new run time record for today
+
+// Ought to keep track of when "yesterday" was last updated and if it was any time "today" then skip this!
+				$queryRunDelete->execute( array($yesterday, $stat->uuid) );	// Remove zero or one rows for yesterday and then insert one row for yesterday.
+				$queryRunInsert->execute( array($stat->uuid, $yesterday, $stat->runTimeHeatYesterday, $stat->runTimeCoolYesterday) );	// Add new run time for yesterday
+			}
+			else
+			{
+				$log->logError( "temps: Error getting run times from {$thermostatRec['id']} {$thermostatRec['tstat_uuid']} {$thermostatRec['ip']} {$thermostatRec['name']}.  No data stored." );
+			}
+
 		}
 		catch( Exception $e )
-		{
+		{	// Does t_lib even throw exceptions?  I don't think it does.
 			$log->logInfo( 'temps: Thermostat Exception: ' . $e->getMessage() );
 		}
 		flock( $lock, LOCK_UN );
@@ -115,6 +139,6 @@ $log->logInfo( "temps: Back from low level communication I have the error code a
 	}
 	fclose($lock);
 }
-$log->logInfo( 'temps: execution time was ' . (microtime(true) - $start_time) . ' seconds.' );
+$log->logInfo( 'temps: End.  Execution time was ' . (microtime(true) - $start_time) . ' seconds.' );
 
 ?>
